% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colbyname.R
\name{colbyname}
\alias{colbyname}
\title{This function receives a list of column names and a data frame,
it then removes or keeps from the df all columns in the list}
\usage{
colbyname(colnames, df, action = "remove")
}
\description{
This function receives a list of column names and a data frame,
it then removes or keeps from the df all columns in the list
}
