% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountNA.R
\name{countNA}
\alias{countNA}
\title{Gives the percentual of NA values of each column ranked by percentual.
For a specific Column simply use a subsetted df.}
\usage{
countNA(data = NULL)
}
\description{
Gives the percentual of NA values of each column ranked by percentual.
For a specific Column simply use a subsetted df.
}
