% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLevels.R
\name{dflevels}
\alias{dflevels}
\title{This function makes a list of all columns levels in a provided
dataframe. For a specific column in a DF, you can easily subset it.}
\usage{
dflevels(data = NULL)
}
\description{
This function makes a list of all columns levels in a provided
dataframe. For a specific column in a DF, you can easily subset it.
}
