% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUsLatLon.R
\name{findUSLatLon}
\alias{findUSLatLon}
\title{This function allows you to find the latitude and longitude based on
US Zip code. A vector can be provided to the function.}
\usage{
findUSLatLon(qzip = NULL)
}
\description{
This function allows you to find the latitude and longitude based on
US Zip code. A vector can be provided to the function.
}
