% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3dlines.R
\name{plot3dlines}
\alias{plot3dlines}
\title{receives 4 vectors of data and a color object to return a 3d interactive
plot with lines. requires package "rgl".
As is, the function works better with the collor.}
\usage{
plot3dlines(xvec, yvec, zvec, colvec = NULL, icolor = topo.colors(50),
  breaknum = 50, ...)
}
\description{
receives 4 vectors of data and a color object to return a 3d interactive
plot with lines. requires package "rgl".
As is, the function works better with the collor.
}
