% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAColumns.R
\name{removeNAColumns}
\alias{removeNAColumns}
\title{receives a DF and remove automatically from it all columns 'that reach
a specified percentual of NA per column}
\usage{
removeNAColumns(df, threshold = 1, savelistnames = F)
}
\description{
receives a DF and remove automatically from it all columns 'that reach
a specified percentual of NA per column
}
