% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetbycollevels.R
\name{subsetbycollevels}
\alias{subsetbycollevels}
\title{Receives a vector of variables to look for, the name of the column
that the variables will be looked upon
the df that will be searched}
\usage{
subsetbycollevels(levelsvec, colname, df)
}
\description{
Receives a vector of variables to look for, the name of the column
that the variables will be looked upon
the df that will be searched
}
