% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUsLatLon.R
\name{findUSLatLon}
\alias{findUSLatLon}
\title{This simple function receives a vector of ZIP codes of any length and
it returns a df with zip, latitude and longitude.}
\usage{
findUSLatLon(qzip = NULL)
}
\description{
This simple function receives a vector of ZIP codes of any length and
it returns a df with zip, latitude and longitude.
}
